🎨 Разбил код на файлы
